學員作品/Web_股價指標圖形整合系統_22周筱蘭/requirements.txt
flask
yfinance
matplotlib.pyplot
matplotlib.backends.backend_tkagg
seaborn